+++ src/thread/thread.cpp
@@ -276,7 +276,7 @@ namespace fc {
      my->yield_until( tp, false );
    }
 
-   int  thread::wait_any_until( std::vector<promise_base::ptr>&& p, const time_point& timeout) {
+   size_t thread::wait_any_until( std::vector<promise_base::ptr>&& p, const time_point& timeout) {
        for( size_t i = 0; i < p.size(); ++i )
          if( p[i]->ready() )
            return i;
@@ -364,12 +364,12 @@ namespace fc {
       return thread::current().exec();
    }
 
-   int wait_any( std::vector<promise_base::ptr>&& v, const microseconds& timeout_us  )
+   size_t wait_any( std::vector<promise_base::ptr>&& v, const microseconds& timeout_us  )
    {
       return thread::current().wait_any_until( fc::move(v), time_point::now() + timeout_us );
    }
 
-   int wait_any_until( std::vector<promise_base::ptr>&& v, const time_point& tp )
+   size_t wait_any_until( std::vector<promise_base::ptr>&& v, const time_point& tp )
    {
       return thread::current().wait_any_until( fc::move(v), tp );
    }
